+++ src/smsfusion/_v2b.py
@@ -7,7 +7,7 @@ from numpy.typing import ArrayLike, NDArray
 from ._ins import _dhda_head, _h_head, _signed_smallest_angle
 from ._transforms import _angular_matrix_from_quaternion as T
 from ._transforms import _euler_from_quaternion, _rot_matrix_from_quaternion
-from ._v2 import (
+from ._v2common import (
     _correct_quat_with_gibbs2,
     _kalman_update_scalar,
     _kalman_update_sequential,
